Refactoring
- move src to mc (library "mc")
- split up unit tests into a couple binaries
- refactor build to be less monolithic
