+++ format.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
-find ./src \( -iname "*.h" -or -iname "*.hpp" -or -iname "*.cpp" \) | xargs clang-format -i -style=file
+find ./mc \( -iname "*.h" -or -iname "*.hpp" -or -iname "*.cpp" \) | xargs clang-format -i -style=file
 find ./tests \( -iname "*.h" -or -iname "*.hpp" -or -iname "*.cpp" \) | xargs clang-format -i -style=file
