+++ mc/meson.build
@@ -1,0 +1,12 @@
+lib_sources = [
+    'marching_cubes.cpp',
+    'storage.cpp',
+    'triangle_soup.cpp',
+    'volume.cpp',
+    'util.cpp'
+]
+
+mc_lib = static_library('mc',
+    sources: lib_sources,
+    include_directories: ['../include'],
+    dependencies: dependencies)
