+++ mc/triangle_soup.cpp
