+++ mc/util.cpp
