+++ tests/meson.build
@@ -1,0 +1,1 @@
+subdir('unit')
