+++ tests/unit/meson.build
@@ -1,0 +1,17 @@
+unit_test_sources = [
+    ['TestSamplers', 'test_samplers.cpp'],
+    ['TestOctree', 'test_octree.cpp'],
+]
+
+# build an exe per test source
+foreach test_source : unit_test_sources
+    test_name = test_source[0]
+    test_source = test_source[1]
+    test_exe = executable(test_name,
+        sources: test_source,
+        include_directories: ['../../include', '../../mc'],
+        link_with: mc_lib,
+        dependencies: dependencies)
+
+    test('unit', test_exe)
+endforeach
